requirements.txt
crawl4ai>=0.4.0
beautifulsoup4>=4.12.0
requests>=2.31.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
faiss-cpu>=1.8.0
tiktoken>=0.7.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
tqdm>=4.66.0
